refactor: 중복 코드 분리
