+++ README.md
@@ -17,8 +17,8 @@
 ### step3
 - [ ] 테마 관리 기능 추가
   - [x] 테마 생성
-  - [ ] 테마 목록 조회
-  - [ ] 테마 삭제
-  - 모든 테마의 시간표는 동일함
+  - [x] 테마 목록 조회
+  - [x] 테마 삭제
+  - [ ] 모든 테마의 시간표는 동일함
 - [ ] 요구사항에 없는 내용이 있다면 직접 요구사항을 추가해서 애플리케이션을 완성
   - [ ] 실제 애플리케이션이라고 생각했을 때 발생할 수 있는 예외 상황을 고려하고 처리한다.
