--- src/main/java/nextstep/reservation/controller/ReservationController.java
@@ -1,41 +1,0 @@
-package nextstep.reservation.controller;
-
-import lombok.RequiredArgsConstructor;
-import nextstep.reservation.dto.ReservationRequestDto;
-import nextstep.reservation.dto.ReservationResponseDto;
-import nextstep.reservation.service.ReservationService;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
-
-@RestController
-@RequestMapping("/reservations")
-@RequiredArgsConstructor
-public class ReservationController {
-    private final ReservationService reservationService;
-
-    @PostMapping
-    public ResponseEntity<Object> addReservation(@RequestBody ReservationRequestDto requestDto) {
-        Long id = reservationService.addReservation(requestDto);
-
-        return ResponseEntity
-                .created(URI.create("/reservations/" + id))
-                .build();
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<ReservationResponseDto> getReservation(@PathVariable Long id) {
-        return ResponseEntity
-                .ok()
-                .body(reservationService.getReservation(id));
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Object> cancelReservation(@PathVariable Long id) {
-        reservationService.deleteReservation(id);
-        return ResponseEntity
-                .noContent()
-                .build();
-    }
-}
