--- src/main/java/nextstep/reservation/dto/ReservationRequestDto.java
@@ -1,28 +1,0 @@
-package nextstep.reservation.dto;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.*;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Builder
-@ToString
-public class ReservationRequestDto {
-    @JsonFormat(pattern = "yyyy-MM-dd")
-    LocalDate date;
-
-    @JsonFormat(pattern = "HH:mm")
-    LocalTime time;
-
-    String name;
-
-    String themeName;
-
-    String themeDesc;
-
-    Integer themePrice;
-}
