--- src/main/java/nextstep/reservation/dto/ReservationResponseDto.java
@@ -1,30 +1,0 @@
-package nextstep.reservation.dto;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.ToString;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-@Getter
-@Builder
-@ToString
-public class ReservationResponseDto {
-    private Long id;
-
-    @JsonFormat(pattern = "yyyy-MM-dd")
-    private LocalDate date;
-
-    @JsonFormat(pattern = "HH:mm")
-    private LocalTime time;
-
-    private String name;
-
-    private String themeName;
-
-    private String themeDesc;
-
-    private Integer themePrice;
-}
