--- src/main/java/nextstep/reservation/dto/ThemeRequestDto.java
@@ -1,16 +1,0 @@
-package nextstep.reservation.dto;
-
-import lombok.*;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Builder
-@ToString
-public class ThemeRequestDto {
-    private String name;
-
-    private String desc;
-
-    private Integer price;
-}
