--- src/main/java/nextstep/reservation/dto/ThemeResponseDto.java
@@ -1,18 +1,0 @@
-package nextstep.reservation.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.ToString;
-
-@Getter
-@Builder
-@ToString
-public class ThemeResponseDto {
-    private Long id;
-
-    private String name;
-
-    private String desc;
-
-    private Integer price;
-}
