--- src/main/java/nextstep/reservation/entity/Theme.java
@@ -1,18 +1,0 @@
-package nextstep.reservation.entity;
-
-import lombok.*;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Builder
-@ToString
-public class Theme {
-    private Long id;
-
-    private String name;
-
-    private String desc;
-
-    private Integer price;
-}
