--- src/main/java/nextstep/reservation/exceptions/exception/DuplicateReservationException.java
@@ -1,4 +1,0 @@
-package nextstep.reservation.exceptions.exception;
-
-public class DuplicateReservationException extends RuntimeException{
-}
