--- src/main/java/nextstep/reservation/exceptions/exception/DuplicateThemeException.java
@@ -1,4 +1,0 @@
-package nextstep.reservation.exceptions.exception;
-
-public class DuplicateThemeException extends RuntimeException{
-}
