+++ src/main/java/nextstep/reservations/domain/entity/theme/Theme.java
@@ -1,0 +1,18 @@
+package nextstep.reservations.domain.entity.theme;
+
+import lombok.*;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Builder
+@ToString
+public class Theme {
+    private Long id;
+
+    private String name;
+
+    private String desc;
+
+    private Integer price;
+}
