+++ src/main/java/nextstep/reservations/dto/reservation/ReservationRequestDto.java
@@ -1,0 +1,28 @@
+package nextstep.reservations.dto.reservation;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.*;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Builder
+@ToString
+public class ReservationRequestDto {
+    @JsonFormat(pattern = "yyyy-MM-dd")
+    LocalDate date;
+
+    @JsonFormat(pattern = "HH:mm")
+    LocalTime time;
+
+    String name;
+
+    String themeName;
+
+    String themeDesc;
+
+    Integer themePrice;
+}
