+++ src/main/java/nextstep/reservations/dto/reservation/ReservationResponseDto.java
@@ -1,0 +1,30 @@
+package nextstep.reservations.dto.reservation;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.ToString;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+@Getter
+@Builder
+@ToString
+public class ReservationResponseDto {
+    private Long id;
+
+    @JsonFormat(pattern = "yyyy-MM-dd")
+    private LocalDate date;
+
+    @JsonFormat(pattern = "HH:mm")
+    private LocalTime time;
+
+    private String name;
+
+    private String themeName;
+
+    private String themeDesc;
+
+    private Integer themePrice;
+}
