+++ src/main/java/nextstep/reservations/dto/theme/ThemeRequestDto.java
@@ -1,0 +1,16 @@
+package nextstep.reservations.dto.theme;
+
+import lombok.*;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Builder
+@ToString
+public class ThemeRequestDto {
+    private String name;
+
+    private String desc;
+
+    private Integer price;
+}
