+++ src/main/java/nextstep/reservations/exceptions/reservation/exception/DuplicateReservationException.java
@@ -1,0 +1,4 @@
+package nextstep.reservations.exceptions.reservation.exception;
+
+public class DuplicateReservationException extends RuntimeException{
+}
