+++ src/main/java/nextstep/reservations/exceptions/theme/exception/DuplicateThemeException.java
@@ -1,0 +1,4 @@
+package nextstep.reservations.exceptions.theme.exception;
+
+public class DuplicateThemeException extends RuntimeException{
+}
